Improve internal step/sink evaluation protocol
We now use dependency injection for step and sink functions so we can
preserve iterators versus sequences for as long as possible. This both
speeds up step processing as well as lets us be much smarter about
handling sinks (e.g., if we still have a sequence by the time `Pipe.nth`
is called, we can directly index instead of slowly iterating).

Closes #5.