Teach Pipe sinks to act as partials with missing source
Previously, Pipe sinks could be called directly as class methods and
used as limited-case partials.

Now, entire Pipe instances can be used as partials by omitting a source;
the final sink call will return a partial in this case, which can be
later passed a source for immediate evaluation of the pipe.